change amounts to double and format according to the currency

+++ src/main/java/com/lemoncash/wallet/movement/operation/MovementOperation.java
@@ -22,7 +22,7 @@ public abstract class MovementOperation {
 
     public abstract Type getType();
 
-    protected Movement generateMovement(Long amount, Wallet wallet) {
+    protected Movement generateMovement(Double amount, Wallet wallet) {
         Movement movement = new Movement();
         movement.setMovementType(getType());
         movement.setAmount(amount);
